// src/packages/bui-react/src/components/_base/Flyout/Flyout.types.ts
type ElementRef = React.RefObject<HTMLElement | null>;

type PassedFlyoutOptions = {
  zIndex?: number | string;
  position?: FlyoutPosition;
  defaultActive?: boolean;
  forcePosition?: boolean;
  keepMounted?: boolean;
  relative?: boolean;
  flyoutArrowElRef: ElementRef;
  flyoutContainerRef?: ElementRef;
  getPositionCompensation?: (
    position: FlyoutPosition,
    trigger: HTMLElement | null,
    arrow: HTMLElement | null
  ) => FlyoutPositionCompensation | undefined;
};

export type FlyoutOptions = Pick<
  PassedFlyoutOptions,
  "forcePosition" | "zIndex"
> & {
  position: FlyoutPosition;
  rtl: boolean;
  mobileFallback?: boolean;
};

export type FlyoutStyles = React.CSSProperties;
export type FlyoutState = {
  styles: FlyoutStyles;
  position: FlyoutPosition;
  active: boolean;
  visible: boolean;
};
type FlyoutMoveAction = {
  type: "move";
  payload: { x: number; y: number; isRTL: boolean };
};
type FlyoutRenderAction = { type: "render"; payload?: never };
type FlyoutShowAction = { type: "show"; payload?: never };
type FlyoutHideAction = { type: "hide"; payload?: never };
type FlyoutRemoveAction = {
  type: "remove";
  payload?: { keepMounted?: boolean };
};
type FlyoutUpdateAction = {
  type: "update";
  payload: Pick<FlyoutState, "styles" | "position">;
};
export type FlyoutAction =
  | FlyoutMoveAction
  | FlyoutRenderAction
  | FlyoutShowAction
  | FlyoutHideAction
  | FlyoutRemoveAction
  | FlyoutUpdateAction;

export type UseFlyout = (
  originRef: ElementRef,
  targetRef: ElementRef,
  options: PassedFlyoutOptions
) => Pick<FlyoutState, "styles" | "position" | "active" | "visible"> & {
  render: () => void;
  show: () => void;
  hide: () => void;
  remove: () => void;
  update: () => void;
  move: (targetX: number, targetY: number) => void;
};